fix review errors
