+++ lib/features/gemini/data/datasources/recipe_data_source.dart
@@ -1,0 +1,4 @@
+abstract class RecipeDataSource {
+  Future<Map<String, dynamic>> fetchRecipe(String dishName);
+  Future<String?> fetchRecipeImage(String dishName);
+}
